\docType{package}
\name{psychobench}
\alias{psychobench}
\alias{psychobench-package}
\title{Supplementary package "(Psycho-)Analysis of Benchmark Experiments"}
\description{
  Supplementary package for the article "(Psycho-)Analysis
  of Benchmark Experiments: A Formal Framework for Investigating the
  Relationship between Data Sets and Learning Algorithms" by
  Eugster, Leisch, and Strobl (see citation("psychobench").
}
